address council review feedback for token tracking
- Make save_crawl_summary atomic with complete_crawl_job (remove
  internal commit, let the next commit cover both)
- Use Decimal instead of float for estimated_cost in schema
- Count api_calls even when usage_metadata is absent
- Type call_details as list[dict[str, Any]]

+++ backend/api/schemas/crawl.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Any
 
 from pydantic import BaseModel, ConfigDict
@@ -32,7 +33,7 @@ class CrawlSummaryResponse(BaseModel):
     input_tokens: int
     output_tokens: int
     thinking_tokens: int
-    estimated_cost: float
+    estimated_cost: Decimal
     created_at: datetime
